requirements.txt
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
openai-whisper
ffmpeg-python
tiktoken
requests
pytest
opencv-python
pytesseract

